pyproject.toml
[tool.poetry]
name = "ecrm-api"
version = "0.1.0"
description = ""
authors = ["migue <miguel.lau@etecsa.cu>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
psycopg2-binary = "^2.9.9"
itsdangerous = "^2.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
FastAPI-SQLAlchemy = "^0.2.1"
python-dotenv = "^1.0.1"
pyi18n-v2 = "^1.2.1"
autopep8 = "^2.2.0"
pydantic = {extras = ["email"], version = "^2.7.3"}
alembic = "^1.13.1"
SQLAlchemy = "^2.0.30"
python-multipart = "^0.0.9"
aiofiles = "^23.2.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

odmantic = "^1.0.2"
httpx = "^0.27.0"
argon2-cffi = "^23.1.0"
argon2-cffi-bindings = "^21.2.0"
pydantic-settings = "^2.3.1"
pytest-asyncio = "^0.23.7"
pymongo = "^4.7.3"
pyjwt = "^2.8.0"
[tool.poetry.dev-dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "ecrm_api.main:start"

[tool.pytest.ini_options]
asyncio_mode = "auto"
